+++ src/clojure/ayler/webapp.clj
@@ -1,13 +1,12 @@
 (ns ayler.webapp
   "The webapp"
-  (:require [ring.util.response :refer (resource-response)]
-            [compojure.route :as route]
+  (:require [ayler.api :as api]
+            [ayler.utils.anti-forgery :refer :all]
+            [compojure.core :refer (defroutes GET context)]
             [compojure.handler :as handler]
-            [ayler.api :as api])
-  (:use [compojure.core :only (defroutes GET context)]
-        [ayler.helpers :only (var-route)]
-        ayler.middleware.anti-forgery-angular
-        ayler.utils.anti-forgery))
+            [compojure.route :as route]
+            [ring.middleware.anti-forgery :refer (wrap-anti-forgery)]
+            [ring.util.response :refer (resource-response)]))
 
 (defn- render
   "convert the output of various enlive functions to string"
@@ -16,14 +15,12 @@
 
 (defroutes app-routes
   (GET "/" _ (anti-forgery-cookie (resource-response "/public/index.html")))
-  (var-route api/app)
+  api/app
   (route/resources "/")
   (route/not-found "NOT FOUND"))
 
-(def pre-app
-  "That's the only way I found to be able to hot-reload the custom middlewares :("
-  (-> (var-route app-routes)
-      wrap-anti-forgery-angular))
-
 (def app
-  (handler/site (var-route pre-app)))
+  (-> app-routes
+      (wrap-anti-forgery {:access-denied-response access-denied-response
+                          :request-token-extractor token-header-extractor})
+      handler/site))