+++ src/javascript/app.js
@@ -5,291 +5,296 @@ $(document).ready(function() {
   $("ul.nav i").tooltip({placement: "bottom"});
 });
 
-angular.module('ayler', [])
-  .config(["$routeProvider", function($routeProvider) {
-    $routeProvider
-      .when("/", {})
-      .when("/:namespace", {
-        controller: NamespaceCtrl,
-        templateUrl: "templates/ns-docstring.html"
-      })
-      .when("/:namespace/:var", {
-        controller: VarInfoCtrl,
-        templateUrl: "templates/var.html"
-      });
-  }]);
+var aylerApp = angular.module("ayler", []);
 
-function MainCtrl($scope) {
-  $scope.setTitle = function(name) {
-    $scope.title = ": " + name;
+aylerApp.config(function($routeProvider) {
+  // Note: the routes use whitespace for template because without
+  // template the controller doesn't run. We don't need the template
+  // itself. We just need the controller to change the state.
+  $routeProvider
+    .when("/",
+          {template: " ",
+           controller: "NsListCtrl"})
+    .when("/:namespace",
+          {template: " ",
+           controller: "NsViewCtrl"})
+    .when("/:namespace/:var",
+          {template: " ",
+           controller: "VarViewCtrl"});
+});
+
+// This will hold all the
+aylerApp.factory("State", function() {
+  var state = {
+    nsList: [],
+    varList: [],
+    errors: [],
+    allNses: [],
+    doc: undefined,   // ns or var docstring
+    source: undefined // var source
   };
 
-  // parse all possibole response status from the server.
-  $scope.handleResponse = function (response, handler) {
-    switch (response.status) {
-    case "disconnected":
-      $scope.$broadcast("connect", {disconnected: true});
-      break;
-    case "not-connected":
-      $scope.$broadcast("connect");
-      break;
-    case "error":
-      $scope.errorHandler(response.response);
-      break;
-    case "done":
-      return handler(response.response);
-    default:
-      alert("Unknown response: " + response);
-    }
-  };
-
-  // Global error handler. Every error in this application should
-  // invoke this method.
-  $scope.errorHandler = function(data, status) {
-    if (status === undefined) {
-      $scope.errors.push(data);
-    } else if (status === 403) { // anti-forgery expired
-      $scope.errors.push("Your session has expired. Please refresh the browser.");
+  // Sets the page title (with constant prefix).
+  state.setTitle = function(text) {
+    if (text) {
+      state.title = "Ayler: " + text;
     } else {
-      $scope.errors.push(data + " (status: " + status + ")");
+      state.title = "Ayler";
     }
-    $scope.anyErrors = true;
   };
 
-  // A common behavior for $http.get.
-  //
-  // params:
-  // * http - $http
-  // * url - The path to get
-  // * flag - A function to set content to ng-show flags.
-  // * handler - The handler for success response
-  $scope.httpFetch = function(http, url, flag, handler) {
-    flag(true);
-    http.get(url)
-      .success(function(data) {
-        flag(false);
-        $scope.handleResponse(data, handler);
-      })
-      .error(function(data, status, headers, config) {
-        flag(false);
-        $scope.errorHandler(data, status);
-      });
-  };
-
-  // Clears the error (e.g. when pressing "Dismiss") so they won't
-  // appear the next time the connection form displays.
-  $scope.clearErrors = function() {
-    $scope.errors = [];
-    $scope.anyErrors = false;
-  };
-
-  $scope.errors = [];
-  $scope.anyErrors = false;
-  $scope.title = "";
-};
-
-function NamespaceListCtrl($scope, $http, $location) {
-  // Listener for the "connect" event to display the connnection
-  // form. Optionally set the disconnected flag which causes a
-  // disconnection error to display.
-  $scope.$on("connect", function(event, args) {
-    if (args && args.disconnected) {
-      $scope.disconnected = true;
-    };
-    $scope.displayConnectForm();
-  });
-
-  $scope.disconnect = function($event) {
-    $event.preventDefault();
-    $http.post("/api/disconnect/")
-      .success(function(data) {
-        $location.path("/");
-        $scope.init();
-      })
-      .error(function(data, status, headers, config) {
-        $scope.errorHandler(data, status);
-      })
-  };
-
-  $scope.setNsLoading = function(b) {$scope.nsLoading = b} // ng-show flag
-  $scope.setVarLoading = function(b) {$scope.varLoading = b} // ng-show flag
-
-  // Display the connect form modal.
-  $scope.displayConnectForm = function() {
-    $("#connectForm").modal('show');
-  };
-
-  // Submit the connection form.
-  $scope.connect = function(){
-    $http.post("/api/remote/", {
-      "port": $scope.remotePort,
-      "host": $scope.remoteHost
-    }).success(function(data) {
-      $("#connectForm").modal('hide');
-      $location.path("/");
-      $scope.init();
-    }).error(function(data, status, headers, config) {
-      $scope.errorHandler(data, status);
+  state.setNsList = function(nslist) {
+    state.nsList = _.map(nslist, function(ns) {
+      return {name: ns, url: escape(ns)};
     });
   };
 
-  // Handler for the namespace-list-response.
-  $scope.namespacesHandler = function(response) {
-    $scope.namespaces = response;
-  };
-
-  // Handler for the var-list-response.
-  $scope.varsHandler = function(response) {
-    $scope.vars = response;
-  };
-
-  // Reset vars filter. Used when selecting new namespace to display.
-  $scope.resetVarsFilter = function() {
-    $scope.vrs = "";
-  };
-
-  // Loads the available namespaces.
-  $scope.loadNamespaces = function() {
-    $scope.httpFetch($http, "api/ls",
-                     $scope.setNsLoading, $scope.namespacesHandler);
-  };
-
-  // Loads available vars in a namespace
-  $scope.loadVars = function(namespace) {
-    $scope.httpFetch($http, "/api/ls/" + namespace,
-                     $scope.setVarLoading, $scope.varsHandler);
-  };
-
-  // Reloads namespace list. Invoked by ngClick.
-  $scope.refreshClicked = function($event) {
-    $event.preventDefault();
-    $scope.loadNamespaces();
-  };
-
-  // Broadcasting message causing the display of the all namespaces
-  // modal.
-  $scope.showAllNsModal = function() {
-    $scope.$broadcast("allNsModal");
-  };
-
-  // Common behavior for execuating on start and on certain refresh
-  // schenarios (e.g. when commiting the connect form).
-  // TODO: Later we should architect this better so we won't need to
-  //       call this function manually.
-  $scope.init = function() {
-    $scope.vars = [] // Initially empty until loadVars() is triggered;
-    $scope.resetVarsFilter();
-    $scope.setNsLoading(false);
-    $scope.setVarLoading(false);
-    $scope.disconnected = false; // ngShow flag
-    $scope.loadNamespaces();
-  };
-
-  $scope.init();
-};
-
-function NamespaceCtrl($scope, $routeParams, $http) {
-  // Handler for displaying namespace docstring.
-  $scope.handleNsDoc = function(response) {
-    $scope.docstring = response || "No Namespace Docs.";
-  };
-
-  $scope.setNsDocLoading = function(b) {$scope.nsDocLoading = b} // ng-show flag
-
-  // Loads the docstring for the namespace.
-  $scope.loadDocstring = function() {
-    var docstringUrl = "/api/doc/" + $scope.nsName;
-    $scope.httpFetch($http, docstringUrl,
-                     $scope.setNsDocLoading, $scope.handleNsDoc);
-  };
-
-  $scope.setNsDocLoading(false);
-  $scope.nsName = escape($routeParams.namespace);
-  $scope.loadVars($scope.nsName);
-  $scope.loadDocstring();
-  $scope.resetVarsFilter();
-  $scope.setTitle($scope.nsName);
-};
-
-function VarInfoCtrl($scope, $routeParams, $http) {
-  // Handler for var docstring.
-  $scope.handleVarDoc = function(response) {
-    $scope.docstring = response || "No Docs.";
-  };
-
-  // Handler for var source.
-  $scope.handleSource = function(response) {
-    if (response) {
-      $scope.source = hljs.highlight("clojure", response).value
-    } else {
-      $scope.source = "<span>Source not found.</span>"
-    }
-  };
-
-  $scope.setDocLoading = function(b) {$scope.docLoading = b} //ng-show flag
-  $scope.setSourceLoading = function(b) {$scope.sourceLoading = b} // ng-show flag
-
-  // Since currently we display either this controller or the
-  // NamespaceCtrl we may need to load the vars on controller
-  // initialization if the user accessed var url directly.
-  $scope.refreshVars = function() {
-    if ($scope.vars.length === 0) {
-      $scope.loadVars($scope.nsName);
+  state.setVarList = function(ns) {
+    return function(varlst) {
+      state.varList = _.map(varlst, function(item) {
+        return {name: item, url: escape(ns + "/" + item)};
+      });
     };
   };
 
-  // Loads the var's docstring.
-  $scope.loadDocstring = function() {
-    var docstringUrl = "/api/doc/" + $scope.nsName + "/" + $scope.varName;
-    $scope.httpFetch($http, docstringUrl,
-                     $scope.setDocLoading, $scope.handleVarDoc);
+  state.setDoc = function(text) {
+    state.doc = text || "No documentation found.";
   };
 
-  // Loads the var's source
-  $scope.loadSource = function() {
-    var sourceUrl = "/api/source/" + $scope.nsName + "/" + $scope.varName;
-    $scope.httpFetch($http, sourceUrl,
-                     $scope.setSourceLoading, $scope.handleSource);
+  state.setSource = function(text) {
+    if (text) {
+      state.source = hljs.highlight("clojure", text).value;
+    } else {
+      state.source = "<span>Source not found.</span>";
+    }
   };
 
-  $scope.nsName = escape($routeParams.namespace);
-  $scope.varName = escape($routeParams.var);
-  $scope.setTitle($scope.nsName + " / " + $scope.varName);
-  $scope.loadDocstring();
-  $scope.loadSource();
-  $scope.refreshVars();
-};
+  state.setAllNses = function(nses) {
+    state.allNses = nses || [];
+    state.selectedNs = _.first(state.allNses);
+  };
 
-function AllNsCtrl($scope, $location, $http) {
-  $scope.$on("allNsModal", function() {
-    $scope.loadAllNses();
-    $("#allNsModal").modal("show");
+  state.appendError = function(error) {
+    state.errors.push(error);
+    state.showErrors = true;
+  };
+
+  state.clearErrors = function() {
+    state.errors = [];
+    state.showErrors = false;
+  };
+
+  return state;
+});
+
+aylerApp.factory("ApiClient", function(State, $http, $rootScope) {
+  var apiClient = {};
+
+  apiClient.handleError = function(data, status) {
+    State.showErrors = true;
+    switch (status) {
+    case undefined:
+      var message =
+        "An unknown error has occured: empty response from the server." +
+        "This may be caused by nrepl connection issues. " +
+        "If it happens again you may need to restart Ayler.";
+      State.appendError(data || message);
+      break;
+    case 403:
+      State.appendError("Your session has expired. Please refresh the browser.");
+      break;
+    default:
+      State.appendError(data + " (status: " + status + ")");
+      break;
+    };
+  };
+
+  // Handles the response according to it's "status" attribute.
+  //
+  // params:
+  // * response: object with "status" and "response" attributes.
+  // * handler: A function which handles the response (e.g. set the
+  //            matching attribute on State) in case the status is
+  //            'done'. Receives the response as sole parameter.
+  apiClient.handleResponse = function(response, handler) {
+    switch(response.status) {
+    case "disconnected":
+      $rootScope.$broadcast("connect", {disconnected: true});
+    case "not-connected":
+      $rootScope.$broadcast("connect");
+    case "done":
+      handler(response.response);
+      break;
+    case "error":
+      apiClient.handleError(response.response);
+      break;
+    default:
+      alert("Unknown response: " + response);
+      break;
+    };
+  };
+
+  // Wrapper for $http.get.
+  //
+  // params:
+  // * url: The url to fetch.
+  // * flag: The attribute on 'State' to set as true while in
+  //         progress.
+  // * handler: A handler to be passsed to 'handleResponse' when
+  //            successfull. See 'handleResponse' for details.
+  apiClient.httpGet = function(url, flag, handler) {
+    State[flag] = true;
+    $http.get(url)
+      .success(function(data) {
+        State[flag] = undefined;
+        apiClient.handleResponse(data, handler);
+      })
+      .error(function(data, status, headers, config) {
+        State[flag] = undefined;
+        apiClient.handleError(data, status);
+      });
+  };
+
+  var defaultErrorHandler = function(data, status, headers, config) {
+    ApiClient.handleError(data, status);
+  }
+
+  // Wrapper for $http.get.
+  //
+  // params:
+  // * url: The url to post to.
+  // * params: The params to post (object);
+  // * succ: A success function. Should accept (data) as parameter.
+  // * err: An error function. Should accept 4 parameters:
+  //        (data, status, headers, config).
+  //        If empty the defaultErrorHandler is used.
+  apiClient.httpPost = function(url, params, succ, err) {
+    $http.post(url, params)
+      .success(succ).error(err || defaultErrorHandler);
+  };
+
+  return apiClient;
+});
+
+aylerApp.directive("errors", function() {
+  return {templateUrl: "templates/errors.html"};
+});
+
+aylerApp.controller("MainCtrl", function($scope, State, ApiClient, $location, $route) {
+  $scope.state = State;
+
+  $scope.$on("connect", function(event, args) {
+    if (args && args.disconnected) {
+      $scope.disconnected = true;
+    } else {
+      $scope.disconnected = false;
+    };
+
+    $("#connectForm").modal("show");
   });
 
-  // handler for all namespaces response
-  $scope.handleAllNses = function(response) {
-    $scope.allNses = response;
-    // avoid null options and reset selection if exists.
-    $scope.selectedNs = response[0];
+  $scope.loadAllNses = function($event) {
+    $event.preventDefault();
+    ApiClient.httpGet("/api/lsall", "allNsBusy", State.setAllNses);
+    $("#allNsModal").modal("show");
   };
 
-  $scope.setAllNsesLoading = function(b) {$scope.allNsesLoading = b}; // ng-show
+  $scope.connect = function() {
+    ApiClient.httpPost(
+      "/api/remote/",
+      {"port": $scope.remotePort,
+       "host": $scope.remoteHost},
+      function(data) {
+        $("#connectForm").modal("hide");
+        $location.path("/");
+        $route.reload();
+      });
+  };
 
-  $scope.loadAllNses = function() {
-    $scope.httpFetch($http, "/api/lsall",
-                     $scope.setAllNsesLoading, $scope.handleAllNses);
+  $scope.disconnect = function($event) {
+    $event.preventDefault();
+    ApiClient.httpPost(
+      "/api/disconnect/", {},
+      function(data) {
+        $route.reload();
+      });
+  };
+
+  $scope.reloadNses = function($event) {
+    $event.preventDefault();
+    ApiClient.httpGet("/api/ls", "nsListBusy", State.setNsList);
   };
 
   $scope.selectNsToRequire = function() {
-    var selected = escape($scope.selectedNs);
+    var selected = escape($scope.state.selectedNs);
     var url = "/api/require/" + selected;
-    $http.post(url, {})
-      .success(function(data) {
+    ApiClient.httpPost(
+      url, {},
+      function(data) {
         $("#allNsModal").modal("hide");
-        $scope.init();
         $location.path("/" + selected);
-      })
-      .error(function(data, status, headers, config) {
-        $scope.errorHandler(data, status);
-      });};
-};
+      });
+  };
+});
+
+// Configure the state for listing namespaces.
+aylerApp.controller("NsListCtrl", function($scope, State, ApiClient) {
+  $scope.state = State;
+  State.varList = [];
+  State.doc = null;
+  State.source = null;
+  State.displayDoc = false;
+  State.displaySource = false;
+  State.symbolName = null; // Pretty name of the queried namespace or var.
+  State.setTitle();
+
+  ApiClient.httpGet("/api/ls", "nsListBusy", State.setNsList);
+});
+
+// Configure the state for displaying namespace.
+aylerApp.controller("NsViewCtrl", function($scope, State, $routeParams, ApiClient) {
+  $scope.state = State;
+  State.source = null;
+  State.displaySource = false;
+  State.displayDoc = true;
+  State.vrs = ""; // clean var filtering on navigation
+  $scope.namespace = unescape($routeParams.namespace);
+  State.symbolName = $scope.namespace;
+  State.setTitle($scope.namespace);
+
+  if (_.isEmpty(State.nsList)) {
+    // only update nsList if it's empty (e.g. landed directly on page)
+    ApiClient.httpGet("/api/ls", "nsListBusy", State.setNsList);
+  }
+  ApiClient.httpGet("/api/ls/" + escape($scope.namespace),
+                    "varListBusy", State.setVarList($scope.namespace));
+  ApiClient.httpGet("/api/doc/" + escape($scope.namespace),
+                    "docBusy", State.setDoc);
+
+});
+
+// Configure the state for displaying var.
+aylerApp.controller("VarViewCtrl", function($scope, State, $routeParams, ApiClient) {
+  $scope.state = State;
+  State.displayDoc = true;
+  State.displaySource = true;
+  $scope.namespace = unescape($routeParams.namespace);
+  $scope.var = unescape($routeParams.var);
+  State.symbolName = $scope.namespace + " / " + $scope.var;
+  State.setTitle(State.symbolName);
+
+  if (_.isEmpty(State.nsList)) {
+    ApiClient.httpGet("/api/ls", "nsListBusy", State.setNsList);
+  };
+
+  if (_.isEmpty(State.varList)) {
+    ApiClient.httpGet("/api/ls/" + escape($scope.namespace),
+                      "varListBusy", State.setVarList($scope.namespace));
+  };
+
+  ApiClient.httpGet("/api/doc/" + escape($scope.namespace + "/" + $scope.var),
+                    "docBusy", State.setDoc);
+
+  ApiClient.httpGet("/api/source/" + escape($scope.namespace + "/" + $scope.var),
+                    "sourceBusy", State.setSource);
+});